Change all references to unsigned int to unsigned long

+++ src/context.h
@@ -13,16 +13,16 @@ class Context {
 public:
 
     Context() {
-        registers = std::vector<unsigned int>(6);
+        registers = std::vector<unsigned long>(6);
         sp = 0;
         pc = 0;
         st = 0;
     }
 
-    Context(std::vector<unsigned int> registers_,
-            unsigned int sp_,
-            unsigned int pc_,
-            unsigned int st_) {
+    Context(std::vector<unsigned long> registers_,
+            unsigned long sp_,
+            unsigned long pc_,
+            unsigned long st_) {
         registers = registers_;
         sp = sp_;
         pc = pc_;
@@ -31,27 +31,27 @@ public:
 
     ~Context() {}
 
-    std::vector<unsigned int> get_registers() {
+    std::vector<unsigned long> get_registers() {
         return registers;
     }
 
-    unsigned int get_sp()  {
+    unsigned long get_sp()  {
         return sp;
     }
 
-    unsigned int get_pc() {
+    unsigned long get_pc() {
         return pc;
     }
 
-    unsigned int get_st() {
+    unsigned long get_st() {
         return st;
     }
 
 private:
-    std::vector<unsigned int> registers;
-    unsigned int sp; // Stack Pointer
-    unsigned int pc; // Program Counter
-    unsigned int st; // Status
+    std::vector<unsigned long> registers;
+    unsigned long sp; // Stack Pointer
+    unsigned long pc; // Program Counter
+    unsigned long st; // Status
 };
 
 #endif // CONTEXT_H
